+++ lib/api/snowplow_events.dart
@@ -441,7 +441,7 @@ class SnowplowTrackerService {
         ),
       );
     } catch (error) {
-      throw error;
+      rethrow;
     }
   }